Separate windows exclusive strings function
