+++ winlogbeat/sys/strings.go
@@ -21,26 +21,7 @@ import (
 	"strings"
 
 	"github.com/elastic/beats/v7/libbeat/common"
-	"golang.org/x/sys/windows"
-	"golang.org/x/text/encoding"
-	"golang.org/x/text/encoding/charmap"
 )
-
-var ansiDecoder *encoding.Decoder
-
-func init() {
-	ansiCP := windows.GetACP()
-	for _, enc := range charmap.All {
-		cm, ok := enc.(*charmap.Charmap)
-		cmID, _ := cm.ID()
-		if !ok || uint32(cmID) != ansiCP {
-			continue
-		}
-		ansiDecoder = cm.NewDecoder()
-		return
-	}
-	ansiDecoder = charmap.Windows1250.NewDecoder()
-}
 
 // UTF16BytesToString converts the given UTF-16 bytes to a string.
 func UTF16BytesToString(b []byte) (string, error) {
@@ -62,11 +43,6 @@ func UTF16BytesToString(b []byte) (string, error) {
 func RemoveWindowsLineEndings(s string) string {
 	s = strings.Replace(s, "\r\n", "\n", -1)
 	return strings.TrimRight(s, "\n")
-}
-
-func ANSIBytesToString(enc []byte) (string, error) {
-	out, err := ansiDecoder.Bytes(enc)
-	return string(out), err
 }
 
 func BinaryToString(bin []byte) string {
