+++ winlogbeat/sys/strings_windows.go
@@ -1,0 +1,45 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package sys
+
+import (
+	"golang.org/x/sys/windows"
+	"golang.org/x/text/encoding"
+	"golang.org/x/text/encoding/charmap"
+)
+
+var ansiDecoder *encoding.Decoder
+
+func init() {
+	ansiCP := windows.GetACP()
+	for _, enc := range charmap.All {
+		cm, ok := enc.(*charmap.Charmap)
+		cmID, _ := cm.ID()
+		if !ok || uint32(cmID) != ansiCP {
+			continue
+		}
+		ansiDecoder = cm.NewDecoder()
+		return
+	}
+	ansiDecoder = charmap.Windows1250.NewDecoder()
+}
+
+func ANSIBytesToString(enc []byte) (string, error) {
+	out, err := ansiDecoder.Bytes(enc)
+	return string(out), err
+}
